Refactor codebase to break circular dependencies
- Moved `VersionMetadata` from `src/storage/version.rs` to `src/core/version.rs` to break `core` <-> `storage` cycle.
- Moved `StorageObserver` and `StorageEvent` from `src/storage/observer.rs` to `src/core/observer.rs` to break `storage` <-> `index` cycle.
- Updated `src/storage/sharding/` to import `TxId` from `crate::core::id` instead of `crate::api` to enforce proper layering (`storage` -> `api` dependency removal).
- Updated all relevant imports across the codebase.
- Verified with `cargo check` and `cargo test`.
