+++ src/api/transaction/write_tx.rs
@@ -18,8 +18,8 @@ use crate::core::id::{EdgeId, IdGenerator, NodeId, VersionId};
 use crate::core::interning::{GLOBAL_INTERNER, InternedString};
 use crate::core::property::{PropertyMap, PropertyMapBuilder};
 use crate::core::temporal::{BiTemporalInterval, Timestamp, time};
+use crate::core::version::VersionMetadata;
 use crate::index::temporal::TemporalIndexes;
-use crate::storage::VersionMetadata;
 use crate::storage::current::CurrentStorage;
 use crate::storage::historical::HistoricalStorage;
 use crate::storage::wal::concurrent_system::ConcurrentWalSystem;
