+++ src/storage/persistence.rs
@@ -12,20 +12,20 @@
 //! - Are created based on time or WAL size thresholds
 //! - Store complete current state + metadata for historical storage
 
-use crate::api::transaction::types::TxId;
 use crate::core::{
     GLOBAL_INTERNER,
+    id::TxId,
     graph::{Edge, Node},
     id::{EdgeId, NodeId, VersionId},
     interning::InternedString,
     property::PropertyMap,
     temporal::{BiTemporalInterval, Timestamp, time},
+    version::VersionMetadata,
 };
 use crate::index::vector::HnswConfig;
 use crate::storage::{
     current::CurrentStorage,
     historical::HistoricalStorage,
-    version::VersionMetadata,
     wal::{LSN, concurrent_system::ConcurrentWalSystem},
 };
 use crate::utils::error::{Result, StorageError};
@@ -1042,8 +1042,8 @@ mod tests {
     use super::*;
     use crate::GLOBAL_INTERNER;
     use crate::core::graph::Node;
-    use crate::storage::version::VersionMetadata;
-    use crate::{PropertyMapBuilder, api::transaction::types::TxId};
+    use crate::core::version::VersionMetadata;
+    use crate::PropertyMapBuilder;
     use tempfile::TempDir;
 
     #[test]
