+++ src/storage/sharding/coordinator.rs
@@ -4,8 +4,7 @@ use super::config::{RebalanceConfig, ShardConfig};
 use super::router::{ShardRouter, TraversalPlan};
 use super::transaction::{DistributedTransaction, DistributedTxError, TwoPhaseCommitLog};
 use super::types::{ShardId, ShardMetrics, ShardState, ShardStatus};
-use crate::api::TxId;
-use crate::core::id::IdGenerator;
+use crate::core::id::{IdGenerator, TxId};
 use std::collections::HashMap;
 use std::sync::{Arc, RwLock};
 use std::time::{Duration, Instant};
