+++ src/storage/sharding/transaction.rs
@@ -4,7 +4,7 @@
 //! shards, ensuring ACID properties across shard boundaries.
 
 use super::types::ShardId;
-use crate::api::TxId;
+use crate::core::id::TxId;
 use std::collections::HashMap;
 use std::fmt;
 use std::sync::atomic::{AtomicU64, Ordering};
